gus_ap_oldi/requirements.txt
Package            Version
------------------ ------------
attrs              23.1.0
certifi            2023.11.17
charset-normalizer 3.3.2
exceptiongroup     1.1.3
idna               3.4
iniconfig          2.0.0
isodate            0.6.1
lxml               4.9.3
packaging          23.2
pip                23.3.1
platformdirs       4.0.0
pluggy             1.3.0
pytest             7.4.3
pytz               2023.3.post1
PyYAML             6.0.1
requests           2.31.0
requests-file      1.5.1
requests-toolbelt  1.0.0
setuptools         58.1.0
six                1.16.0
tomli              2.0.1
urllib3            2.1.0
wheel              0.41.3
xmltodict          0.13.0
zeep               4.2.1
